docs/requirements.txt
matplotlib
sphinx_design
sphinx_copybutton
sphinx-gallery
scikit-learn
pydata-sphinx-theme
numpydoc